% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsh_soil_properties.R
\name{dsh_soil_properties}
\alias{dsh_soil_properties}
\title{'Dynamic Soils Hub' Soil Property Grids}
\usage{
dsh_soil_properties(
  x = NULL,
  variables = NULL,
  resolutions = NULL,
  aggregations = NULL,
  top_depth = 0,
  bottom_depth = 200,
  region = "conus",
  subregion = "conus",
  filename = NULL,
  overwrite = FALSE,
  vrt = FALSE,
  ...
)
}
\arguments{
\item{x}{An R spatial object (such as a \emph{SpatVector}, \emph{SpatRaster}, or \emph{sf}
object). Default: \code{NULL} returns a virtual raster. If \code{x} is a \emph{SpatRaster}
the coordinate reference system, extent, and resolution are used as a
template for the output raster.}

\item{variables}{\emph{character}. One or more variables corresponding to grid
file names (without .tif extension). See
\url{https://s3-fpac-nrcs-dshub-public.s3.us-east-1.amazonaws.com/SoilProperties/_README.txt}
for details.}

\item{resolutions}{integer. One or more of: \code{30}, \code{900}. Use \code{NULL} for no
filter on resolution.}

\item{aggregations}{\emph{character}. One or more of: \code{"domcond"}, \code{"wtdavg"},
\code{"minmax"}. Use \code{NULL} for no filter on aggregation method.}

\item{top_depth}{integer. Top depth. Default: \code{0} (includes all grids that
have no depth)}

\item{bottom_depth}{integer. Bottom depth. Default: \code{200}}

\item{region}{\emph{character}. One or more of: \code{"conus"}, \code{"oconus"}. Use \code{NULL}
for no filter on location.}

\item{subregion}{\emph{character}. One or more of: \code{"conus"}, \code{"pr"}, \code{"ak"},
\code{"as"}, \code{"fm"}, \code{"gu"}, \code{"hi"}, \code{"mh"}, \code{"mp"}, \code{"pw"}}

\item{filename}{character. Path to write output raster file. Default: \code{NULL}
will keep result in memory (or store in temporary file if memory threshold
is exceeded)}

\item{overwrite}{logical. Overwrite \code{filename} if it exists? Default: \code{FALSE}}

\item{vrt}{logical. Use \code{terra::vrt()} for result SpatRaster? Default:
\code{FALSE}}

\item{...}{Additional arguments passed to \code{writeRaster()} via \code{terra::crop()}
or \code{terra::project()}. Ignored when \code{x} is \code{NULL}.}
}
\value{
A \emph{SpatRaster} object.
}
\description{
Create a 'terra' \emph{SpatRaster} object referencing soil property grids from the
\href{https://s3-fpac-nrcs-dshub-public.s3.us-east-1.amazonaws.com/SoilProperties/_README.txt}{'Dynamic Soils Hub' Public S3 bucket}
}
\examples{
\dontshow{if (requireNamespace("terra") && as.logical(Sys.getenv("R_RDSHUB_RUN_LONG_EXAMPLES", unset = "FALSE"))) withAutoprint(\{ # examplesIf}
library(rdshub)
library(terra)

x <- dsh_soil_properties(
  # variables = c("claytotal", "sandtotal", "silttotal", "drainagecl"),
  top_depth = 0,
  bottom_depth = c(0, 200),
  aggregations = c("wtdavg", "domcond"),
  resolutions = 900
)

terra::sources(x)

terra::metags(x, layer = 1)

vars <- x |>
  terra::metags(seq_len(terra::nlyr(x))) |>
  subset(name == "variable")
names(x) <- vars$value

terra::plot(x)

terra::plot(x$sandtotal + x$claytotal + x$silttotal)

y <- dsh_soil_properties(variables = "mukey",
                         aggregation = "domcond",
                         bottom_depth = 0)
y
terra::plot(y[2])
\dontshow{\}) # examplesIf}
}
