% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdshub-package.R
\docType{package}
\name{rdshub-package}
\alias{rdshub}
\alias{rdshub-package}
\title{rdshub: Functions for the 'Dynamic Soils Hub'}
\description{
Experimental functions for interacting with public-facing 'Dynamic Soils Hub' tools and APIs.
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov} (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})

}
\keyword{internal}
