% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdshub-package.R
\name{dsh_s3_base_url}
\alias{dsh_s3_base_url}
\title{DSHub Sources}
\usage{
dsh_s3_base_url()
}
\value{
\code{dsh_s3_base_url()}: \emph{character}. URL for public DSHub S3 bucket. When set,
the value of R option \code{rdshub.s3_base_url} is returned instead of the
hard-coded default.
}
\description{
Various methods for returning URLs for DSHub sources. Each function returns
value of user-customizable R options, with a fallback to the current public
default value.
}
\examples{
# URL for public DSHub S3 bucket
dsh_s3_base_url()
}
